+++ training/src/pages/Login/Login.jsx
@@ -1,6 +1,5 @@
 /* eslint-disable */
 import React from 'react';
-import * as yup from 'yup';
 import PropTypes from 'prop-types';
 import { Redirect } from 'react-router-dom';
 import {
@@ -10,28 +9,9 @@ import {
 import { Email, VisibilityOff, LockOutlined } from '@material-ui/icons';
 import callApi from '../../libs/utils/api';
 import { MyContext } from '../../contexts/snackBarProvider/index';
+import schema from './schema';
+import Design from './style';
 
-const schema = yup.object().shape({
-  email: yup.string()
-    .trim().email().required('Email Address is a required field'),
-  password: yup.string()
-    .required('Password is required'),
-});
-const Design = (theme) => ({
-  icon: {
-    background: 'red',
-    marginLeft: theme.spacing(22),
-    marginTop: theme.spacing(2),
-  },
-  main: {
-    width: 400,
-    marginTop: theme.spacing(20),
-    marginLeft: theme.spacing(58),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-});
 class Login extends React.Component {
   constructor(props) {
     super(props);
@@ -69,8 +49,7 @@ class Login extends React.Component {
      const response = await callApi(data, 'post', '/user/login');
       this.setState({ loading: false });
       if (response.status === 200) {
-        const a = localStorage.setItem('token', response.data);
-        console.log('a :', a);
+       localStorage.setItem('token', response.data);
         this.setState({
           redirect: true,
           hasError: false,
