+++ training/src/pages/Login/schema.js
@@ -1,0 +1,9 @@
+import * as yup from 'yup';
+
+const schema = yup.object().shape({
+  email: yup.string()
+    .trim().email().required('Email Address is a required field'),
+  password: yup.string()
+    .required('Password is required'),
+});
+export default schema;
