+++ training/src/pages/Login/style.js
@@ -1,0 +1,16 @@
+const Design = (theme) => ({
+  icon: {
+    background: 'red',
+    marginLeft: theme.spacing(22),
+    marginTop: theme.spacing(2),
+  },
+  main: {
+    width: 400,
+    marginTop: theme.spacing(20),
+    marginLeft: theme.spacing(58),
+  },
+  submit: {
+    margin: theme.spacing(3, 0, 2),
+  },
+});
+export default Design;
